Recycle the ValidatedMessage type to allow for decoding messages while retaining metadata
